requirements.txt
flask
flask-cors
requests
python-dotenv
PyPDF2
beautifulsoup4
werkzeug
